regular size of form control
By default this version of bootstrap adds a 14px font-size to form elements. It also hard codes a size. On iOS small form elements gets zoomed in until they're over 16px. That means that for a size difference that you can barely see, a zoom out needs to be done after entering the date.

fixes #230